% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_profile_functions.R
\name{align_soil_layers}
\alias{align_soil_layers}
\title{Align soil profile data to a new set of depths to base of layer}
\usage{
align_soil_layers(new_SLB, profile, drop_deeper_layers = TRUE)
}
\arguments{
\item{new_SLB}{a numeric vector with new values for depth to base of layer (SLB)
to which soil profile data should be aligned}

\item{profile}{a tibble containing soil profile data including a column for depth
to base of layer (SLB)}

\item{drop_deeper_layers}{a logical value indicating whether layers below the maximum
value in `new_SLB` should be dropped. If TRUE the values will be dropped (default),
if FALSE layers will be kept.}
}
\description{
Align soil profile data to a new set of depths to base of layer
}
\examples{

align_soil_layers(new_SLB, profile)

}
