% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_functions.R
\name{filter_mukey_by_coord}
\alias{filter_mukey_by_coord}
\title{Filter data associated with SSURGO or STATSGO map unit keys by latitude and longitude}
\usage{
filter_mukey_by_coord(mukey_tbl, pt_geom, soil_name)
}
\arguments{
\item{mukey_tbl}{a numeric or character vector of map unit keys}

\item{lat}{a numeric value of latitude}

\item{long}{a numeric value of longitude}
}
\description{
Filter data associated with SSURGO or STATSGO map unit keys by latitude and longitude
}
