% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_functions.R
\name{pull_profile_by_coords}
\alias{pull_profile_by_coords}
\title{Pull STATSGO or SSURGO soil profile by coordinates}
\usage{
pull_profile_by_coords(
  pt_geom = NULL,
  lat = NULL,
  long = NULL,
  site = "",
  SSURGO = TRUE,
  STATSGO = !SSURGO
)
}
\arguments{
\item{pt_geom}{an sf object which contains the coordinates
(latitude and longitude) of the point for which soil profile
data are being pulled}

\item{lat}{a numeric value of the decimal degree latitude
of the point for which soil profile data are being pulled}

\item{long}{a numeric value of the decimal degree longitude
of the point for which soil profile data are being pulled}

\item{site}{the site name for which the soil profile is being pulled}

\item{SSURGO}{logical argument indicating whether SSURGO profiles
should be included in query}

\item{STATSGO}{logical argument indicating whether STATSGO profiles
should be included in query}
}
\description{
Pull STATSGO or SSURGO soil profile by coordinates
}
