% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SSURGO_functions.R
\name{pull_profile_by_name}
\alias{pull_profile_by_name}
\title{Pull STATSGO or SSURGO soil profile by soil series name}
\usage{
pull_profile_by_name(
  soil_name,
  state = "",
  county = "",
  pt_geom = NULL,
  lat = NULL,
  long = NULL,
  SSURGO = TRUE,
  STATSGO = !SSURGO
)
}
\arguments{
\item{state}{a character vector corresponding to `soil_names` that gives
the state of the location from which the profile should be queried}

\item{county}{a character vector corresponding to `soil_names` that gives
the county of the location from which the profile should be queried}

\item{pt_geom}{an sf object which contains the coordinates
(latitude and longitude) of the point for which soil profile
data are being pulled}

\item{lat}{a numeric value of the decimal degree latitude
of the point for which soil profile data are being pulled}

\item{long}{a numeric value of the decimal degree longitude
of the point for which soil profile data are being pulled}

\item{SSURGO}{logical argument indicating whether SSURGO profiles
should be included in query}

\item{STATSGO}{logical argument indicating whether STATSGO profiles
should be included in query}

\item{soil_names}{a character vector of soil series names to query}
}
\description{
Pull STATSGO or SSURGO soil profile by soil series name
}
