% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_align_depths.R
\name{soil_align_depths}
\alias{soil_align_depths}
\title{Align soil profile data to a new set of depths to base of layer}
\usage{
soil_align_depths(
  soil_data,
  new_depths,
  depth_name = "SLB",
  drop_deeper_layers = TRUE
)
}
\arguments{
\item{soil_data}{a data frame containing soil profile data including a column
for depth of base of layer}

\item{new_depths}{a numeric vector with new values for depth to base of layer
to which soil profile data should be aligned}

\item{depth_name}{a character string containing the name for the column in
\code{soil_data} that contains the depth of base of layer}

\item{drop_deeper_layers}{a logical value indicating whether layers below the maximum
value in \code{new_depths} should be dropped. If TRUE the values will be dropped (default),
if FALSE layers will be kept.}
}
\description{
Align soil profile data to a new set of depths to base of layer
}
\examples{

soil_align_depths(soil_depth, new_depths)

}
