% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_ptf_hwsd_sldr.R
\name{soil_ptf_hwsd_sldr}
\alias{soil_ptf_hwsd_sldr}
\title{Determine soil drainage based on WRB number}
\usage{
soil_ptf_hwsd_sldr(wrb_number)
}
\arguments{
\item{wrb_number}{the world resources based (WRB) number for soil types}
}
\description{
Determine soil drainage parameter (SLDR) based on the drainage
 class as defined in the Harmonized World Soil Database (HWSD)
 for a given world resources based (WRB) number
}
