% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_ptf_slu1.R
\name{soil_ptf_slu1}
\alias{soil_ptf_slu1}
\title{Determine stage 1 soil evaporation limit}
\usage{
soil_ptf_slu1(silt, clay, depth)
}
\arguments{
\item{silt}{a vector or SpatRaster with percent silt content}

\item{clay}{a vector or SpatRaster with percent clay content}

\item{depth}{a vector or SpatRaster with dimensions corresponding to silt and
clay with the depth to the base of soil layer}
}
\description{
Determine the stage 1 soil evaporation limit following Ritchie et al. (1989)
 based on sand, silt and clay percentage in the first 15-cm of the soil
 profile.
}
