% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_saxton_slll.R
\name{soil_saxton_slll}
\alias{soil_saxton_slll}
\title{Saturated conductivity based on Saxton and Rawls (2006)}
\usage{
soil_saxton_slll(silt, clay, soc, bulk_density, coarse_fraction)
}
\arguments{
\item{silt}{silt percentage of soil (percent)}

\item{clay}{clay percentage of soil (percent)}

\item{soc}{soil organic carbon of soil (percent)}

\item{bulk_density}{soil bulk density (g cm^{-3})}

\item{coarse_fraction}{volumetric fraction of coarse  (> 2mm diameter)
fragments (cm^3 cm^{-3})}
}
\description{
Estimates saturated conductivity based on the pedotransfer functions
defined in Saxton and Rawls (2006).
}
