% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_saxton_ssks.R
\name{soil_saxton_ssks}
\alias{soil_saxton_ssks}
\title{Saturated conductivity based on Saxton and Rawls (2006)}
\usage{
soil_saxton_ssks(theta_s, theta_33, theta_1500, coarse_fraction, bulk_density)
}
\arguments{
\item{theta_s}{volumetric water at saturation (cm^3 cm^{-3})}

\item{theta_33}{volumetric water at -33 kPa (cm^3 cm^{-3})}

\item{theta_1500}{volumetric water at -1500 kPa (cm^3 cm^{-3})}

\item{coarse_fraction}{volumetric fraction of coarse  (> 2mm diameter)
fragments (cm^3 cm^{-3})}

\item{bulk_density}{soil bulk density (g cm^{-3})}
}
\description{
Estimates saturated conductivity based on the pedotransfer functions
defined in Saxton and Rawls (2006).
}
