% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_ptf_nrcs_hsg.R
\name{soil_ptf_nrcs_hsg}
\alias{soil_ptf_nrcs_hsg}
\title{Determine hydrologic soil group}
\usage{
soil_ptf_nrcs_hsg(ssks, depth)
}
\description{
Determine hydrologic soil group (HSG) based on saturated conductivity
 with depth and the method described in Chapter 7 - "Hydrologic Soil
  Groups" of the NRCS Part 630 National Engineering Handbook (2009)
}
