% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soil_ptf_slu1.R
\name{soil_ptf_slu1}
\alias{soil_ptf_slu1}
\title{Determine stage 1 soil evaporation limit}
\usage{
soil_ptf_slu1(fc, pwp, depth)
}
\arguments{
\item{fc}{a vector or SpatRaster with field capacity values as volumetric
soil water in units of cm^3 cm^-3}

\item{pwp}{a vector or SpatRaster with permanent wilting point values as
volumetric soil water in units of cm^3 cm^-3}

\item{depth}{a vector or SpatRaster with dimensions corresponding to fc and
pwp with the depth to the base of soil layer}
}
\description{
Determine the stage 1 soil evaporation limit based on field capacity,
 permanent wilting point and soil depth
}
